# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: RamoVerde API
  version: 1.0.0
  description: Versioned REST contract shared by the RamoVerde web and mobile clients.
paths:
  /api/v1/auth/csrf:
    get:
      operationId: getAuthCsrf
      tags:
      - v1
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfToken'
          description: ''
...
  /api/v1/auth/password/change:
    post:
      operationId: postAuthPasswordChange
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordChange'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordChange'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordChangeResponse'
          description: 'Mobile: new token pair.'
        '204':
          description: 'Web: session kept, other sessions revoked.'
  /api/v1/auth/password/reset:
    post:
      operationId: postAuthPasswordReset
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      security:
      - {}
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: ''
  /api/v1/auth/password/reset/confirm:
    post:
      operationId: postAuthPasswordResetConfirm
      description: 'Sets the new password. Does not sign in: MFA is required at next
        sign-in.'
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
        required: true
      security:
      - {}
      responses:
        '204':
          description: No response body
  /api/v1/auth/session/login:
    post:
      operationId: postAuthSessionLogin
      description: 'Password step. Does not sign in: a second factor must follow.'
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionLogin'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SessionLogin'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SessionLogin'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionAuthFlow'
          description: ''
  /api/v1/auth/session/logout:
    post:
      operationId: postAuthSessionLogout
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/auth/sessions:
    get:
      operationId: getAuthSessions
      tags:
      - v1
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSession'
          description: ''
  /api/v1/auth/sessions/{id}:
    delete:
      operationId: deleteAuthSession
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/auth/sessions/revoke-all:
    post:
      operationId: postAuthSessionsRevokeAll
      description: 'Global logout: every web session and every mobile refresh token,
        this one included.'
      tags:
      - v1
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/auth/sessions/revoke-others:
    post:
      operationId: postAuthSessionsRevokeOthers
      tags:
      - v1
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/auth/token:
    post:
      operationId: postAuthToken
      description: 'Password step. Never returns tokens: complete it with a second
        factor.'
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileCredentials'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MobileCredentials'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MobileCredentials'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAuthFlow'
          description: ''
  /api/v1/auth/token/logout:
    post:
      operationId: postAuthTokenLogout
      description: Unauthenticated auth endpoint that reports failures as 401 with
        a generic message.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileLogout'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MobileLogout'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MobileLogout'
        required: true
      security:
      - {}
      responses:
        '204':
          description: No response body
...
  /api/v1/auth/token/refresh:
    post:
      operationId: postAuthTokenRefresh
      description: Unauthenticated auth endpoint that reports failures as 401 with
        a generic message.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileRefresh'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MobileRefresh'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MobileRefresh'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenResponse'
          description: ''
...
  /api/v1/users/me:
    get:
      operationId: getUsersMe
      tags:
      - v1
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
components:
  schemas:
    AuthFlowStatusEnum:
      enum:
      - mfa_required
      - mfa_enrollment_required
      - authenticated
      type: string
      description: |-
        * `mfa_required` - Second factor required
        * `mfa_enrollment_required` - Second factor enrollment required
        * `authenticated` - Authenticated
    CsrfToken:
      type: object
      properties:
        csrfToken:
          type: string
      required:
      - csrfToken
    Detail:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
    KindEnum:
      enum:
      - web
      - mobile
      type: string
      description: |-
        * `web` - Web
        * `mobile` - Mobile
    MfaMethodEnum:
      enum:
      - totp
      - recovery
      - webauthn
      type: string
      description: |-
        * `totp` - TOTP
        * `recovery` - Recovery code
        * `webauthn` - WebAuthn
...
    MobileAuthFlow:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/AuthFlowStatusEnum'
        methods:
          type: array
          items:
            $ref: '#/components/schemas/MfaMethodEnum'
        challenge:
          type: string
          description: Short-lived, single-use MFA challenge.
      required:
      - challenge
      - methods
      - status
...
    MobileCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
          maxLength: 4096
      required:
      - email
      - password
    MobileLogout:
      type: object
      properties:
        refresh:
          type: string
          writeOnly: true
      required:
      - refresh
...
    MobileRefresh:
      type: object
      properties:
        refresh:
          type: string
      required:
      - refresh
    MobileTokenResponse:
      type: object
      properties:
        access:
          type: string
        refresh:
          type: string
        recovery_codes:
          type: array
          items:
            type: string
          description: 'Only after first MFA enrollment: show once, never stored by
            the app.'
      required:
      - access
      - refresh
...
    PasswordChange:
      type: object
      properties:
        current_password:
          type: string
          writeOnly: true
          maxLength: 4096
        new_password:
          type: string
          writeOnly: true
          maxLength: 4096
      required:
      - current_password
      - new_password
    PasswordChangeResponse:
      type: object
      description: 'Returned only to mobile sessions: their token pair is re-issued.'
      properties:
        access:
          type: string
        refresh:
          type: string
      required:
      - access
      - refresh
    PasswordResetConfirm:
      type: object
      properties:
        uid:
          type: string
          writeOnly: true
          maxLength: 64
        token:
          type: string
          writeOnly: true
          maxLength: 128
        new_password:
          type: string
          writeOnly: true
          maxLength: 4096
      required:
      - new_password
      - token
      - uid
    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
      - email
...
    SessionAuthFlow:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/AuthFlowStatusEnum'
        methods:
          type: array
          items:
            $ref: '#/components/schemas/MfaMethodEnum'
        user:
          $ref: '#/components/schemas/User'
        recovery_codes:
          type: array
          items:
            type: string
      required:
      - methods
      - status
    SessionLogin:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
          maxLength: 4096
      required:
      - email
      - password
...
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        first_name:
          type: string
          readOnly: true
          title: Nome
        last_name:
          type: string
          readOnly: true
          title: Cognome
        password_change_required:
          type: boolean
          readOnly: true
        password_expires_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - email
      - first_name
      - id
      - last_name
      - password_change_required
      - password_expires_at
    UserSession:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        kind:
          allOf:
          - $ref: '#/components/schemas/KindEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        last_seen_at:
          type: string
          format: date-time
          readOnly: true
        ip_address:
          type: string
          readOnly: true
          nullable: true
        user_agent:
          type: string
          readOnly: true
        mfa_method:
          type: string
          readOnly: true
        current:
          type: boolean
          readOnly: true
      required:
      - created_at
      - current
      - id
      - ip_address
      - kind
      - last_seen_at
      - mfa_method
      - user_agent
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: http://localhost:8010