# packages/api-client/src/generated/model/detail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * RamoVerde API
 * Versioned REST contract shared by the RamoVerde web and mobile clients.
 * OpenAPI spec version: 1.0.0
 */

export interface Detail {
  detail: string;
}