# packages/api-client/src/generated/model/kindEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * RamoVerde API
 * Versioned REST contract shared by the RamoVerde web and mobile clients.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `web` - Web
 * * `mobile` - Mobile
 */
export type KindEnum = (typeof KindEnum)[keyof typeof KindEnum];

export const KindEnum = {
  web: 'web',
  mobile: 'mobile',
} as const;