# packages/api-client/src/generated/model/passwordChange.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * RamoVerde API
 * Versioned REST contract shared by the RamoVerde web and mobile clients.
 * OpenAPI spec version: 1.0.0
 */

export interface PasswordChange {
  /** @maxLength 4096 */
  current_password: string;
  /** @maxLength 4096 */
  new_password: string;
}