# packages/api-client/src/generated/model/passwordChangeResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * RamoVerde API
 * Versioned REST contract shared by the RamoVerde web and mobile clients.
 * OpenAPI spec version: 1.0.0
 */

/**
 * Returned only to mobile sessions: their token pair is re-issued.
 */
export interface PasswordChangeResponse {
  access: string;
  refresh: string;
}