# packages/api-client/src/generated/model/passwordResetConfirm.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * RamoVerde API
 * Versioned REST contract shared by the RamoVerde web and mobile clients.
 * OpenAPI spec version: 1.0.0
 */

export interface PasswordResetConfirm {
  /** @maxLength 64 */
  uid: string;
  /** @maxLength 128 */
  token: string;
  /** @maxLength 4096 */
  new_password: string;
}